+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   * Add from\_api method to ApiAccess
   * Add from\_api method to DueBadges
   * Add from\_api method to Evening and Evening::Activity
+  * Add from\_api method to Event
 
 ## Version 0.0.11